+++ src/solverpy/benchmark/launcher.py
@@ -2,12 +2,13 @@ import os
 import random
 import logging
 
-from ..benchmark.path import bids
+from .path import bids
 from ..tools import human, log
 from ..task.solvertask import SolverTask
 from ..task.bar import SolvingBar, RunningBar
 from ..task import launcher 
 from .reports import markdown
+from .db.providers.solved import Solved
 
 logger = logging.getLogger(__name__)
 
@@ -32,11 +33,33 @@ def init(setup=None):
 def jobname(solver, bid, sid):
    return f"{solver}:{sid} @ {bid}"
 
-def run(solver, bid, sid, desc=None, taskdone=None, db=None, cores=4, shuffle=True, force=False, **others):
+def run(solver, bid, sid, desc=None, taskdone=None, db=None, cores=4, shuffle=True, force=False, solvedby=None, it=None, **others):
+   others = dict(others, force=force, it=it, solvedby=solvedby)
    desc = desc if desc else jobname(solver, bid, sid)
    logger.debug(f"evaluating {desc}: {jobname(solver, bid, sid)}")
    # prepare the tasks to be evaluated
    ps = bids.problems(bid)
+   skipped = {}
+   if solvedby and (it == 0):
+      solvable = Solved(bid, solvedby, solver.limits.limit).cache
+      if solvable:
+         simulate = dict(status="TIMEOUT", runtime=solver.limits.timeout, limit=solver.limits.limit)
+         skipped = {p:dict(simulate) for p in (set(ps) - solvable)}
+         if taskdone:
+            for (p,res) in skipped.items():
+               taskdone(solver.solved(res))
+         if db:
+            tasks = [SolverTask(solver,bid,sid,p) for p in skipped]
+            results = [simulate]*len(tasks)
+            db.store(tasks, results)
+         ps = solvable
+         logger.debug(f"evaluation: restricted to {len(solvable)} problems solvable by {solvedby}")
+   #solvable = None
+   #if solvedby and (it == 0):
+   #   solvable = Solved(bid, solvedby, solver.limits.limit).cache
+   #   if solvable:
+   #      logger.debug(f"evaluation: restricted to {len(solvable)} problems solvable by {solvedby}")
+   #ps = solvable if solvable else bids.problems(bid) 
    tasks = [SolverTask(solver,bid,sid,p) for p in ps]
    logger.debug(f"evaluation: {len(tasks)} tasks scheduled")
    # check for the (cached) results in the database
@@ -66,6 +89,7 @@ def run(solver, bid, sid, desc=None, taskdone=None, db=None, cores=4, shuffle=Tr
       # compose the cached and new results
       results = {t.problem:res for (t,res) in zip(todo, results)}
       results.update(done)
+      results.update(skipped)
       logger.debug(f"evaluation done: +{bar._solved} -{bar._unsolved} !{bar._errors}")
    else:
       logger.debug("evaluation skipped: already done")